feat: moved function to util file
